+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.2 (01/01/2018)
+
+- fixes spacing with nested values that have longer names than the parent
+- ensure the height of the matrix is the largest height possible
+
 # 0.0.1 (01/01/2018)
 
 - basic usage documented and tested